Fix some bugs

+++ Src/AMAK.Application/Services/Options/OptionService.cs
@@ -50,7 +50,14 @@ namespace AMAK.Application.Services.Options {
             var existingOption = await _optionRepository.GetAll()
                     .FirstOrDefaultAsync(x => x.Id == id && x.ProductId == productId && !x.IsDeleted) ?? throw new NotFoundException("Option not found!");
 
-            _optionRepository.Remove(existingOption);
+            if (existingOption.IsDeleted) {
+                _optionRepository.Remove(existingOption);
+            }
+
+            existingOption.IsDeleted = true;
+            existingOption.DeleteAt = DateTime.UtcNow;
+
+            _optionRepository.Update(existingOption);
 
             await _optionRepository.SaveChangesAsync();
